split view component
