+++ src/views/Search.js
@@ -3,10 +3,12 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Helmet } from 'react-helmet';
 import { withRouter } from 'react-router-dom';
-import { Header, Table, Icon } from 'semantic-ui-react';
+import { Header } from 'semantic-ui-react';
 
 import TitleHeader from '../components/TitleHeader';
 import SearchMap from '../components/SearchMap';
+
+import SearchResult from './SearchResult';
 
 import strings from '../resources/strings';
 
@@ -47,34 +49,7 @@ class Search extends Component {
 		}
 	}
 
-	renderLists (data) {
-		this.key = data.key;
-		this.name = data.name;
-		this.distance = data.distance;
-		this.rentDiscountPrice = data.rentDiscountPrice;
-		this.stayDiscountPrice = data.stayDiscountPrice;
-
-		return (
-			<Table.Row key={this.key}>
-				<Table.Cell textAlign='center'>
-					<Icon name='h' />{this.name}
-				</Table.Cell>
-				<Table.Cell textAlign='center'>
-					{this.distance}
-				</Table.Cell>
-				<Table.Cell textAlign='center'>
-					{parseInt(this.rentDiscountPrice, 10).toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',')}
-				</Table.Cell>
-				<Table.Cell textAlign='center'>
-					{parseInt(this.stayDiscountPrice, 10).toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',')}
-				</Table.Cell>
-			</Table.Row>
-		);
-	}
-
 	render () {
-		const { search } = this.props;
-
 		return (
 			<div>
 				<Helmet>
@@ -92,44 +67,15 @@ class Search extends Component {
 					markerPositionLng={defaultMarkerPositionLng}
 					onMarkerDragEnd={this.onMarkerDragEnd}
 				/>
-				<div>
-					<Table celled selectable>
-						<Table.Header>
-							<Table.Row>
-								<Table.HeaderCell>
-									<Header as='h4' textAlign='center'>{strings.accommodationName}</Header>
-								</Table.HeaderCell>
-								<Table.HeaderCell>
-									<Header as='h4' textAlign='center'>{strings.distance}</Header>
-								</Table.HeaderCell>
-								<Table.HeaderCell>
-									<Header as='h4' textAlign='center'>{strings.rentReservation}</Header>
-								</Table.HeaderCell>
-								<Table.HeaderCell>
-									<Header as='h4' textAlign='center'>{strings.stayReservation}</Header>
-								</Table.HeaderCell>
-							</Table.Row>
-						</Table.Header>
-						<Table.Body>
-							{
-								search && search.map(this.renderLists, this)
-							}
-						</Table.Body>
-					</Table>
-				</div>
+				<SearchResult />
 			</div>
 		);
 	}
 }
 
 Search.propTypes = {
-	searchAction: PropTypes.func.isRequired,
-	search: PropTypes.array.isRequired
+	searchAction: PropTypes.func.isRequired
 };
-
-const mapStateToProps = state => ({
-	search: state.search
-});
 
 const mapDispatchToProps = dispatch => ({
 	searchAction (date, lat, lng) {
@@ -138,6 +84,6 @@ const mapDispatchToProps = dispatch => ({
 });
 
 export default withRouter(connect(
-	mapStateToProps,
+	null,
 	mapDispatchToProps
 )(Search));
